+++ profileService/internal/domain/program/service.go
@@ -4,6 +4,6 @@ import "context"
 
 type Service interface {
 	GetProgramNamesByInstanceID(ctx context.Context, courseID int64) ([]*string, error)
-	GetProgramNameByID(ctx context.Context, id int) (*string, error)
+	GetProgramNameByID(ctx context.Context, id int64) (*string, error)
 	GetAllPrograms(ctx context.Context) ([]*Program, error)
 }
